Move pub permissions job files to rdl_to_prj
Since these jobs will run on a fairly frequent basis we are moving them from one_off to a new folder: rdl_to_prj. In addition we needed to make changes to the gitlab ci file to deploy these files successfully. There was also an update to the time function in those scripts since to_timestamp is not available in spark 2.1.0.